A twisty maze of hacks ...
Various changes from testing against mongooseim and to unfuckify the
halfass removal of global state from the previous change.

The (very old) ejabberd in ubuntu 14.04 appears to use the old muc
protocol. With the new muc protocol, rooms are created in a locked
state. Hackily accomodate this by always sending a "muc#owner" protocol
set stanza, even if the room exists because there doesn't appear to be a
way of getting response the response from the xmpp library.

Change some of the global state to parameters passed to functions which
broke temporary rooms.

Now requires a modified copy of go-xmpp, since the message body may
contain other elements, breaking the hack for checking for delayed
messages.

And to rant a bit: the way go encourages both the use of global state
and data hiding instead of composable, pure functions makes the language
and it's libraries really frustrating to use.

Next step: re-write!